feat(ci): unified pipeline with cross-compilation, E2E VM testing, and init systems
Unified CI pipeline (ci.yml):
- 32 jobs: lint, test, build, packages, Docker E2E, VM E2E
- ARM64 native builds on ubuntu-24.04-arm runners
- Cross-compilation for arm, 386, riscv64 (glibc + musl via Zig CC)
- Native BSD builds on Proxmox VMs (FreeBSD, OpenBSD, NetBSD)
- macOS builds (amd64 Intel + arm64 Apple Silicon)
- 12 E2E VM tests (9 Linux distros + 3 BSDs)
- Package verification for .deb, .rpm, .apk, .pkg.tar.zst, .xbps

E2E hardening:
- Section-defensive validate-probe.sh with omitempty null guards
- Thermal zones, top_processes null safety for all platforms
- FreeBSD package plist generation for pkg create
- Binary verification after package install in install.sh
- Robust jq install with functional verification and static fallback
- SCP retry logic (3 attempts) for VM resilience

Init system support (10 systems):
- systemd, OpenRC, runit, s6, dinit, SysVinit
- FreeBSD rc.d, OpenBSD rc.d, NetBSD rc.d, macOS launchd

Also includes: MkDocs documentation, API proto definitions,
configuration examples, and devcontainer updates.

+++ src/lib/probe/crates/probe-quota/src/freebsd.rs
@@ -198,32 +198,44 @@ fn read_rlimits_into(limits: &mut QuotaLimits) {
         // RLIMIT_NOFILE (if not already set from rctl)
         if limits.nofile_limit.is_none() {
             if getrlimit(RLIMIT_NOFILE, &mut rl) == 0 {
-                limits.nofile_limit =
-                    Some(if rl.rlim_cur == libc::RLIM_INFINITY { u64::MAX } else { rl.rlim_cur });
+                limits.nofile_limit = Some(if rl.rlim_cur == libc::RLIM_INFINITY {
+                    u64::MAX
+                } else {
+                    rl.rlim_cur as u64
+                });
             }
         }
 
         // RLIMIT_CPU (if not already set)
         if limits.cpu_time_limit_secs.is_none() {
             if getrlimit(RLIMIT_CPU, &mut rl) == 0 {
-                limits.cpu_time_limit_secs =
-                    Some(if rl.rlim_cur == libc::RLIM_INFINITY { u64::MAX } else { rl.rlim_cur });
+                limits.cpu_time_limit_secs = Some(if rl.rlim_cur == libc::RLIM_INFINITY {
+                    u64::MAX
+                } else {
+                    rl.rlim_cur as u64
+                });
             }
         }
 
         // RLIMIT_DATA (if not already set)
         if limits.data_limit_bytes.is_none() {
             if getrlimit(RLIMIT_DATA, &mut rl) == 0 {
-                limits.data_limit_bytes =
-                    Some(if rl.rlim_cur == libc::RLIM_INFINITY { u64::MAX } else { rl.rlim_cur });
+                limits.data_limit_bytes = Some(if rl.rlim_cur == libc::RLIM_INFINITY {
+                    u64::MAX
+                } else {
+                    rl.rlim_cur as u64
+                });
             }
         }
 
         // RLIMIT_NPROC (if not already set)
         if limits.pids_limit.is_none() {
             if getrlimit(RLIMIT_NPROC, &mut rl) == 0 {
-                limits.pids_limit =
-                    Some(if rl.rlim_cur == libc::RLIM_INFINITY { u64::MAX } else { rl.rlim_cur });
+                limits.pids_limit = Some(if rl.rlim_cur == libc::RLIM_INFINITY {
+                    u64::MAX
+                } else {
+                    rl.rlim_cur as u64
+                });
             }
         }
     }
